refractoring and order service

+++ ecom-order-service/src/main/java/com/ecom/order/repository/OrderRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ecom.order.domain.Orders;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Orders, Long>{
+public interface OrderRepository extends JpaRepository<Orders, Integer>{
 
 }
